+++ internal/constants/constants.go
@@ -14,10 +14,10 @@ package constants
 
 const (
 	// SdkVersion is the version of the OpenFGA Go SDK.
-	SdkVersion = "0.7.4"
+	SdkVersion = "0.7.5"
 
 	// UserAgent is the user agent used in HTTP requests.
-	UserAgent = "openfga-sdk go/0.7.4"
+	UserAgent = "openfga-sdk go/0.7.5"
 
 	// SampleBaseDomain is the example API domain for documentation/tests.
 	SampleBaseDomain = "fga.example"