+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v0.7.5
 
 ### [0.7.5](https://github.com/openfga/go-sdk/compare/v0.7.4...v0.7.5) (2026-02-23)
+
 - feat: replace errgroup with conc pool in WriteExecute method (#249) - thanks @AltuisticIsopod
 
 ## v0.7.4